修复另外一个线程占有该对象的问题

+++ src/Acorisoft.Studio/ViewModels/StickyNotes/StickyNoteGalleryViewModel.cs
@@ -34,11 +34,11 @@ namespace Acorisoft.Studio.ViewModels
 
             //
             // 按创建时间排序 按修改时间排序 
-            NewCommand = ReactiveCommand.Create(OnNewItem, _engine.IsOpen);
-            DeleteThisCommand = ReactiveCommand.Create<StickyNoteIndexWrapper>(OnDeleteThis, _engine.IsOpen);
-            DeleteAllCommand = ReactiveCommand.Create(OnDeleteAll, _engine.IsOpen);
-            DeleteThisPageCommand = ReactiveCommand.Create(OnDeleteThisPage, _engine.IsOpen);
-            OpenThisCommand = ReactiveCommand.Create<StickyNoteIndexWrapper>(OnOpenItem, _engine.IsOpen);
+            NewCommand = ReactiveCommand.Create(OnNewItem, _css.IsOpen);
+            DeleteThisCommand = ReactiveCommand.Create<StickyNoteIndexWrapper>(OnDeleteThis, _css.IsOpen);
+            DeleteAllCommand = ReactiveCommand.Create(OnDeleteAll, _css.IsOpen);
+            DeleteThisPageCommand = ReactiveCommand.Create(OnDeleteThisPage, _css.IsOpen);
+            OpenThisCommand = ReactiveCommand.Create<StickyNoteIndexWrapper>(OnOpenItem, _css.IsOpen);
         }
 
         public async Task SearchAsync(string keyword)
@@ -84,12 +84,22 @@ namespace Acorisoft.Studio.ViewModels
 
         protected async void OnDeleteThisPage()
         {
-            await _engine.DeleteThisPageAsync();
+            if (await ViewAware.AwaitDelete("是否全部删除", "删除操作将会清空所有内容") == true)
+            {
+                //
+                // 等待创建
+                await _engine.DeleteThisPageAsync();
+            }
         }
 
         protected async void OnDeleteThis(StickyNoteIndexWrapper item)
         {
-            await _engine.DeleteThisAsync(item.Source);
+            if (await ViewAware.AwaitDelete("是否全部删除", "删除操作将会清空所有内容") == true)
+            {
+                //
+                // 等待创建
+                await _engine.DeleteThisAsync(item.Source);
+            }
 
             //
             // 跳转
